move tests from test_ignore.py to the right place
